template/python3-flask-debian/requirements.txt
flask
waitress
tox==3.*
pillow
minio